Revert back to original implementation
