+++ src/seedu/addressbook/commands/ViewCommand.java
@@ -2,7 +2,7 @@ package seedu.addressbook.commands;
 
 import seedu.addressbook.common.Messages;
 import seedu.addressbook.data.person.ReadOnlyPerson;
-import seedu.addressbook.data.person.Person;
+import seedu.addressbook.ui.TextUi;
 
 
 /**
@@ -33,7 +33,7 @@ public class ViewCommand extends Command {
             if (!addressBook.containsPerson(target)) {
                 return new CommandResult(Messages.MESSAGE_PERSON_NOT_IN_ADDRESSBOOK);
             }
-            String targetDetails = Person.getPrintableString(target.getPhone(), target.getEmail(), target.getAddress());
+            String targetDetails = TextUi.getPrintableString(target.getPhone(), target.getEmail(), target.getAddress());
             return new CommandResult(String.format(MESSAGE_VIEW_PERSON_DETAILS, target.getAsTextHidePrivate(targetDetails)));
         } catch (IndexOutOfBoundsException ie) {
             return new CommandResult(Messages.MESSAGE_INVALID_PERSON_DISPLAYED_INDEX);
