+++ src/seedu/addressbook/ui/TextUi.java
@@ -2,8 +2,8 @@ package seedu.addressbook.ui;
 
 import static seedu.addressbook.common.Messages.*;
 import seedu.addressbook.commands.CommandResult;
+import seedu.addressbook.data.person.Printable;
 import seedu.addressbook.data.person.ReadOnlyPerson;
-import seedu.addressbook.data.person.Person;
 
 import java.io.InputStream;
 import java.io.PrintStream;
@@ -135,7 +135,7 @@ public class TextUi {
     private void showPersonListView(List<? extends ReadOnlyPerson> persons) {
         final List<String> formattedPersons = new ArrayList<>();
         for (ReadOnlyPerson person : persons) {
-        	String personDetails = Person.getPrintableString(person.getName(), person.getPhone(), person.getEmail(), person.getAddress());
+        	String personDetails = getPrintableString(person.getName(), person.getPhone(), person.getEmail(), person.getAddress());
         	formattedPersons.add(person.getAsTextHidePrivate(personDetails));
         }
         showToUserAsIndexedList(formattedPersons);
@@ -165,4 +165,15 @@ public class TextUi {
     private static String getIndexedListItem(int visibleIndex, String listItem) {
         return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
     }
+    
+    /**
+     * Returns a concatenated version of the printable strings of each object.
+     */
+	public static  String getPrintableString(Printable... printables) {
+		String concatString = "";
+		for (Printable p : printables) {
+			concatString += p.getPrintableString() + " ";
+        }
+		return concatString.trim();
+	}
 }
